+++ hexital/core/candle.py
@@ -120,6 +120,41 @@ class Candle:
     def high_low(self) -> float:
         return abs(self.high - self.low)
 
+    def as_list(self) -> list:
+        """
+        Generates a list of values from the OHLCV values.
+        `[timestamp, open, high, low, close, volume]` in that order.
+        With an optional `timedelta` value at the end being the `timeframe`
+
+        Returns:
+            list: A list of the `Candle` values
+        """
+        return [self.timestamp, self.open, self.high, self.low, self.close, self.volume] + (
+            [self.timeframe] if self.timeframe else []
+        )
+
+    def as_dict(self) -> dict:
+        """
+        Generates a dict of values from the OHLCV values, with  the following keys:
+        `[timestamp, open, high, low, close, volume]`
+
+        Returns:
+            dict: A list of the `Candle` values
+        """
+        cdl = {
+            "open": self.open,
+            "high": self.high,
+            "low": self.low,
+            "close": self.close,
+            "volume": self.volume,
+            "timestamp": self.timestamp,
+        }
+
+        if self.timeframe:
+            cdl["timeframe"] = self.timeframe
+
+        return cdl
+
     @classmethod
     def from_dict(cls, candle: Dict[str, Any]) -> Candle:
         """