+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 ## 2.0.1
 
-*Release Date: -*
+*Release Date: 2025-01-12**
 
 - Fixes
     - Corrected merged candles not combining aggregation factor